+++ src/pagination/SizePerPageDropDown.js
@@ -14,18 +14,32 @@ class SizePerPageDropDown extends Component {
       className,
       variation,
       btnContextual,
+      isBootstrap4,
       currSizePerPage
     } = this.props;
 
+    if (hidden) return null;
+
     const openClass = open ? 'open show' : '';
-    const dropDownStyle = { visibility: hidden ? 'hidden' : 'visible' };
+
+    const renderOptions = () => {
+      const attrs = {
+        className: `dropdown-menu ${openClass}`,
+        role: 'menu',
+        'aria-labelledby': 'pageDropDown'
+      };
+      const type = isBootstrap4 ? 'div' : 'ul';
+
+      return React.createElement(type, attrs, options);
+    };
 
     return (
-      <span style={ dropDownStyle }
+      <span
         className={ `${variation} ${openClass} ${className} ${sizePerPageDefaultClass}` }>
         <button className={ `btn ${btnContextual} dropdown-toggle` }
           id='pageDropDown' data-toggle='dropdown'
           aria-expanded={ open }
+          aria-haspopup={ !open }
           onClick={ onClick }
           onBlur={ onBlur }>
           { currSizePerPage }
@@ -34,9 +48,7 @@ class SizePerPageDropDown extends Component {
           <span className='caret'/>
           </span>
         </button>
-        <ul className='dropdown-menu' role='menu' aria-labelledby='pageDropDown'>
-          { options }
-        </ul>
+        { renderOptions() }
       </span>
     );
   }